collectors: cut the sockshop-specific processing to targets.sockshop module
